+++ EnergyOptimizer.Service/DTOs/DashboardUpdateDto.cs
@@ -3,7 +3,7 @@
     public class DashboardUpdateDto
     {
         public DateTime Timestamp { get; set; }
-        public double TotalConsumption { get; set; }
+        public decimal TotalConsumption { get; set; }
         public int ActiveDevices { get; set; }
         public int TotalReadings { get; set; }
         public List<TopConsumerDto> TopConsumers { get; set; } = new();
@@ -13,6 +13,6 @@
     {
         public int DeviceId { get; set; }
         public string DeviceName { get; set; } = string.Empty;
-        public double CurrentConsumption { get; set; }
+        public decimal CurrentConsumption { get; set; }
     }
 }
